Changed Entity logic

+++ Data/Migrations/ApplicationDbContextModelSnapshot.cs
@@ -23,11 +23,11 @@ namespace Data.Migrations
 
             modelBuilder.Entity("Domain.Entities.Articles.Article", b =>
                 {
-                    b.Property<int>("Id")
+                    b.Property<long>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("int");
+                        .HasColumnType("bigint");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"));
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("Id"));
 
                     b.Property<string>("Message")
                         .HasColumnType("nvarchar(max)");
